added docs for grafana/prometheus steps
Added minimum necessary steps for users to access Grafana/Prometheus for metrics
and visualization.

updated prometheus manifests, set default grafana vars for local use

Updated the prometheus manifests by copying the newer version of the file from
the original source. I also changed a couple of the default values to reflect
the recommendations made in the Grafana/Prometheus steps in the readme, which
make it easier to access Grafana locally (by removing authentication; users can
add this if they need it).

uncommented prometheus-manifests in init.sh for ipfs-cluster

added grafana instructions to end of init.sh files

added prometheus-manifests.yml to ipfs-cluster

The Prometheus manifests will be updated to account for ipfs-cluster related
scraping/metrics-gathering. In lieu of that, and to make everything consistent
(`prometheus-manifests.yml` in same place as `go-ipfs-deployment.yml`), a copy
is now in the `ipfs-cluster` folder. In the future, may want to simply call the
`prometheus-manifests.yml` from the root directory and extend those directly for
consistency.