added docs for grafana/prometheus steps
Added minimum necessary steps for users to access Grafana/Prometheus for metrics
and visualization.

updated prometheus manifests, set default grafana vars for local use

Updated the prometheus manifests by copying the newer version of the file from
the original source. I also changed a couple of the default values to reflect
the recommendations made in the Grafana/Prometheus steps in the readme, which
make it easier to access Grafana locally (by removing authentication; users can
add this if they need it).