Refactor commit to complete PR Added a pre-run validation function to exit the test immediately if any step has a common error in its selection structure.  Errors include running steps with multiple selections, including selections that run on too many or nonexistant nodes, and trying to run on subset indices without specifying a partition
Restructured code to bubble up error in go style rather than terminate with fatal.  Note that partitions
are also robust to yaml mistakes and such mistakes will also exit the program before running any steps.

Restructured test folder, and added a script to run tests on the selection feature.  This commit adds
tests to ensure that bad yaml files fail before running any steps.  As with the other meta tests in
the test_test folder they are not used exactly as intended, in this case because no assertions are
reached.  The testing script will output success and failure messages based on whether the tests
return an error or success code.