Param search/replace reworked; cleaned up options handling
The bulk of this commit, involves a heavy rewrite of the search/replace
mechanism for parameters both in its implementation and its error handling. The
test file is searched for all (unique) parameter identifiers (of the form
`{{ <paramName> }}`, see below), each of which is then compared against the
parameter name -> value replacement mappings specified by the user via the
`--param` flag. If the parameter name is not found in the replacement map, an
error is thrown. If a parameter replacement in the map is never referenced in a
test (i.e. the test doesn't include an identifier corresp. to that parameter's
name), that parameter replacement is simply ignored.

One smaller change involves the parameter declarations, which are now of the
form `{{ <paramName> }}`, with 0 or more spaces before/after `<paramName>`.
Unfortunately, the current implementation does not have a syntax to ignore
double-braces (e.g. `\{{ <stuff> }}`), as Go's regex engine does not support
negative lookbehind and the easiest workaround (that I've thougth of, at least)
would unecessarily complicate the code. However, all double-braced strings that
do not strictly meet the parameter syntax are ignored, and thus may be used (to
some extent) independently of this feature (e.g. strings in the test file of the
form `{{\ <stuff> }}` will be left alone).

Further, parameter name identifiers must begin with an alphabetical character or
underscore, then may be followed by zero or more alphanumeric + underscore
characters. There are currently no restrictions on the form of a parameter's
replacement value.

Rebased commit: adding params.go

Rebased commit: added params file feature

A file may be used to input a test config that, so far, only supports a
`params` entry for parameters.

Example config:

```config.yml
params:
    N: 3
    M: 10
```

By default, the program will look for this file in `<testdir>/config.yml`, where
`<testdir>` is the directory containing the test that is being run. If this file
does not exist, a warning will be reported but the test will continue. An
alternative file may be specified via the `--config` flag. If this flag is used
and the input file is not found, the test will fail.